chore(application): setup webpack to build successfully
ensure that webpack is setup, builds correctly, copies the necessary files to the dist folder, loads images and works with jsx

[delivers #152720959]
