+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 .nyc_output/
 package-lock.json
 coverage
+dist/
