+++ client/index.html
@@ -1,27 +1,24 @@
 <!DOCTYPE html>
 <html lang="en">
-    <head>
-        <title></title>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
 
-        <!-- Google Font -->
-        <link href="https://fonts.googleapis.com/css?family=Great+Vibes" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css?family=Encode+Sans" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css?family=Ubuntu:300" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css?family=Titillium+Web:300" rel="stylesheet">
-        <link href="https://fonts.googleapis.com/css?family=Quicksand" rel="stylesheet">
+<head>
+  <title></title>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <!-- Google Font -->
+  <link href="https://fonts.googleapis.com/css?family=Berkshire+Swash|Cairo|Maven+Pro" rel="stylesheet">
 
-        <!-- CSS  -->
-        <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.100.2/css/materialize.min.css">
-        <!-- <link rel="stylesheet" href="assets/css/styles.css"> -->
-        
-    </head>
-    <body>
-        <div id="application">
-            
-        </div>
-    <script src="bundle.js"></script>
-    </body>
+  <!-- CSS  -->
+  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+  <link rel="stylesheet" href="assets/css/aos.css">
+  <link href="assets/css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection" />
+  <link href="assets/css/style.css" type="text/css" rel="stylesheet" media="screen,projection" />
+
+</head>
+
+<body>
+  <div id="root"></div>
+  <script src="/bundle.js"></script>
+</body>
+
 </html>
