+++ package.json
@@ -3,14 +3,14 @@
   "version": "1.0.0",
   "description": "Hello Books Application",
   "scripts": {
-    "prestart": "babel-node tools/startMessage.js",
-    "start": "npm-run-all --parallel lint:watch start:server",
-    "open:client": "babel-node tools/clientServer.js",
-    "lint": "node_modules/.bin/esw webpack.config.* client tools server",
-    "lint:server": "node_modules/.bin/esw server",
-    "lint:watch": "npm run lint -- --watch",
+    "start": "npm-run-all --parallel test:watch lint:watch open:client",
     "start:server": "NODE_ENV=development babel-node  ./server/bin/www.js --exec babel-node --presets env ",
     "test": "NODE_ENV=test nyc --reporter=text --reporter=html mocha test/server --timeout 150000 --compilers js:babel-register",
+    "open:client": "webpack -d && npm run copy:assets && babel-node tools/clientServer.js ",
+    "copy:assets": "cp -r ./client/assets ./dist/. && cp ./client/index.html ./dist/index.html",
+    "lint": "node_modules/.bin/esw webpack.config.* client tools server",
+    "lint:watch": "npm run lint -- --watch",
+    "test:watch": "npm run test -- --watch",
     "seed:db": "NODE_ENV=test babel-node server/scripts/initialize.js",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
