+++ webpack.config.babel.js
@@ -1,0 +1,55 @@
+import webpack from 'webpack';
+import path from 'path';
+
+export default {
+  devtool: 'inline-source-map',
+  entry: [
+    'eventsource-polyfill',
+    'webpack-hot-middleware/client?reload=true',
+    path.resolve(__dirname, './client/index.jsx')
+  ],
+  target: 'web',
+  output: {
+    path: path.join(__dirname, 'dist'),
+    filename: 'bundle.js'
+  },
+  devServer: {
+    contentBase: path.resolve(__dirname, 'client')
+  },
+  plugins: [new webpack.HotModuleReplacementPlugin(), new webpack.NoEmitOnErrorsPlugin()],
+  resolve: {
+    extensions: ['.js', '.jsx']
+  },
+  module: {
+    loaders: [{
+        test: /\.js(x)$/,
+        include: path.join(__dirname, 'client'),
+        loaders: ['babel-loader']
+      },
+      {
+        test: /(\.css)$/,
+        loaders: ['style-loader', 'css-loader']
+      },
+      {
+        test: /\.(jpg|png|jpeg|svg)$/,
+        loader: 'url-loader?name=[name].[ext]&limit=10',
+      },
+      {
+        test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'file-loader'
+      },
+      {
+        test: /\.(woff|woff2)$/,
+        loader: 'url-loader?prefix=font/&limit=5000'
+      },
+      {
+        test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'url-loader?limit=10000&mimetype=application/octet-stream'
+      },
+      {
+        test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
+        loader: 'url-loader?limit=10000&mimetype=image/svg+xml'
+      }
+    ]
+  }
+};
